cambios final

+++ Backend/API_PESO_PIG/Models/Stage.Model.cs
@@ -5,10 +5,10 @@ namespace API_PESO_PIG.Models
     public class Stage
     {
         [Key]
-        public int id_Stage {  get; set; }
+        public int id_Stage { get; set; }
         public string Name_Stage { get; set; }
-        public int Weight_From {  get; set; }
-        public int Weight_Upto { get; set; }
+        public float Weight_From { get; set; }
+        public float Weight_Upto { get; set; }
         public int Dur_Stage { get; set; }
     }
 }
